Fix for new SwerveControllerFollower() constructor
We added a new SwerveControllerFollower() constructor which can form a trajectory from a PathPlanner file, and also set the initial pose. But there was a bug. This should fix that bug, though it is untested at this time.

Existing code uses new SwerveControllerFollower(...).beforeStarting(() -> driveBase.resetOdometry(initial)).

The issue is that while we are building the initial pose at the same time as the existing code (when the command is constructed), we were mistakenly trying to also apply that pose at the same time! This fix moves setting the pose to when the command is actually executed, first thing in initialize(), which should match the semantics of the existing code.